requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
youtube-transcript-api==0.6.1
openai==1.63.2
python-dotenv==1.0.0
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
requests==2.31.0
psycopg[binary]==3.2.4
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
alembic==1.13.1
